DEVEXP-498 Can now define default metadata
I realized my initial "DocumentBatch" design was not good because it limited the user to defining a single set of default metadata. And the v1/documents endpoint allows for many sets of default metadata. The new design - introducing a DefaultMetadata class, and allowing the user to provide a list of any number of Document and DefaultMetadata instances - is simpler and adheres to what /v1/documents allows too. 

No docs yet. Going to do a separate PR for those next.