Moved eval guts into impl/eval.py
Also reworked `client.eval` so it accepts args for the scripts.

The thought behind `impl` is that since everything is public in Python, "impl" gives us a way of communicating that whatever is in here isn't intended to be used by a user. So out modules not in "impl" can access stuff in "impl", but users should not.