Returning None instead of empty array
This seems like the right choices for v1/rows, where a JSON object is returned by default. For eval, an "empty sequence" is different from an "empty array" - an empty array is something that still exists (an array with no items in it) - and so when an empty sequence is returned for eval/invoke, None seems more appropriate too.